backend/requirements.txt
fastapi
uvicorn
python-multipart
neo4j<6.0.0
openai
langchain
langchain-community
langchain-openai
langchain-google-genai
pypdf
python-docx
tiktoken
python-dotenv
requests
networkx
